Changes:
Army: added method to calculate army's combat value.  Added method to get army's leader.
Skill: Changed skill effects from int to double, to remove need to convert to double later.
Character: added method to calculate character's combat value (for a military engagement).  Added method to calculate character's leadership value (if army leader).  Made minor changes to account for changes to Skill.

+++ hist_mmorpg/Skill.cs
@@ -21,7 +21,7 @@ namespace hist_mmorpg
         /// <summary>
         /// Holds skill effects
         /// </summary>
-        public Dictionary<string, int> effects;
+        public Dictionary<string, double> effects;
 
         /// <summary>
         /// Constructor for Skill
@@ -29,7 +29,7 @@ namespace hist_mmorpg
 		/// <param name="id">String holding skill ID</param>
 		/// <param name="nam">String holding skill name</param>
         /// <param name="effs">Dictionary<string name, int effect> holding skill effects</param>
-		public Skill(String id, String nam, Dictionary<string, int> effs)
+        public Skill(String id, String nam, Dictionary<string, double> effs)
         {
 
             // validate nam length = 1-20
